src/rca-demo/pyproject.toml
[tool.poetry]
name = "rca-demo"
version = "0.1.0"
description = ""
authors = ["Dimitar Gueorguiev <dimitar_pg13@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
typing-extensions = "^4.7.1"
networkx = "^3.1"
pyspark = "^3.4.1"
pandas = "^2.0.3"
numpy = "^1.25.1"
"ruamel.yaml" = "^0.17.32"

[tool.poetry.dev-dependencies]
pytest = "^7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
